Rename "calendars" to "calendar"
