+++ src/main/java/seedu/address/logic/commands/calendar/ShowCommand.java
@@ -1,8 +1,8 @@
-package seedu.address.logic.commands.calendars;
+package seedu.address.logic.commands.calendar;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.calendars.CliSyntax.PREFIX_MONTH;
-import static seedu.address.logic.parser.calendars.CliSyntax.PREFIX_YEAR;
+import static seedu.address.logic.parser.calendar.CliSyntax.PREFIX_MONTH;
+import static seedu.address.logic.parser.calendar.CliSyntax.PREFIX_YEAR;
 
 import java.util.GregorianCalendar;
 
@@ -21,12 +21,8 @@ public class ShowCommand extends Command {
 
     public static final String MESSAGE_SUCCESS = "Showed calendar";
     public static final String MESSAGE_USAGE = getCommandFormat(COMMAND_WORD) + ": Shows the calendar view.\n"
-            + "Parameters: "
-            + PREFIX_YEAR + "YEAR "
-            + PREFIX_MONTH + "MONTH" + "\n"
-            + "Example: " + getCommandFormat(COMMAND_WORD) + " "
-            + PREFIX_YEAR + "2018 "
-            + PREFIX_MONTH + "12";
+            + "Parameters: " + PREFIX_YEAR + "YEAR " + PREFIX_MONTH + "MONTH" + "\n" + "Example: "
+            + getCommandFormat(COMMAND_WORD) + " " + PREFIX_YEAR + "2018 " + PREFIX_MONTH + "12";
 
     private final Index year;
     private final Index month;
