+++ src/main/java/seedu/address/logic/parser/calendar/ShowCommandParser.java
@@ -1,13 +1,13 @@
-package seedu.address.logic.parser.calendars;
+package seedu.address.logic.parser.calendar;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.calendars.CliSyntax.PREFIX_MONTH;
-import static seedu.address.logic.parser.calendars.CliSyntax.PREFIX_YEAR;
+import static seedu.address.logic.parser.calendar.CliSyntax.PREFIX_MONTH;
+import static seedu.address.logic.parser.calendar.CliSyntax.PREFIX_YEAR;
 
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.calendars.ShowCommand;
+import seedu.address.logic.commands.calendar.ShowCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
